CRUD

+++ src/main/java/springCrud/dao/UserDao.java
@@ -8,11 +8,11 @@ public interface UserDao {
 
     void add(User user);
 
-    void updateUser(int id, User user);
+    void updateUser(Long id, User user);
 
-    public void deleteUser(int id);
+    public void deleteUser(Long id);
 
-    public User getUserById(int id);
+    public User getUserById(Long id);
 
     public List<User> getAllUsers();
